+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## 1.3.1 - 2026-01-26
 
 - Bumped CUDA from 12.6 to 12.8 (adds Blackwell/RTX 50 series support)
+- Added GTX 10 series compatibility note to README
+- Log Flash Attention availability at startup
 
 ## 1.3.0 - 2026-01-26